feat: Separate the visual styles of chat bubbles for users and the assistant, and clarify where the JSON parsing strategy is implemented
